+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format mostly follows [Keep a Changelog](http://keepachangelog.com/en/1.0.0/
 
 - New command-line option `--prepare-jobs` to initialize new jobs or jobs without history (#831 by nille02)
 - New reporter: `ntfy` (#854 by fyrk)
+- New `json` option for `url` jobs: when set to `true`, serializes `data` as JSON and sets `Content-type: application/json`
 
 ### Changed